fix : backend dev 모드로 변경, testInitData(테스트모드 : 주문, 결제, 배송), BaseInitData(공통 : member, 상품, 결제 옵션) 분리
