+++ backend/src/main/resources/application-test.yml
@@ -8,3 +8,9 @@ spring:
     console:
       enabled: true
       path: /h2-console
+logging:
+  level:
+    org.hibernate.orm.jdbc.bind: TRACE
+    org.hibernate.orm.jdbc.extract: TRACE
+    org.springframework.transaction.interceptor: TRACE
+    com.back: DEBUG
