+++ backend/src/main/resources/application.yml
@@ -2,7 +2,7 @@ spring:
   application:
     name: caffe
   profiles:
-    active: test
+    active: dev
   output:
     ansi:
       enabled: always
@@ -37,10 +37,3 @@ spring:
 
 springdoc:
   default-produces-media-type: application/json;charset=UTF-8
-
-logging:
-  level:
-    org.hibernate.orm.jdbc.bind: TRACE
-    org.hibernate.orm.jdbc.extract: TRACE
-    org.springframework.transaction.interceptor: TRACE
-    com.back: DEBUG
